Bug 560393, part 2: (xpcom directory) Fix compile warning: format $f expects type $t1, but argument 2 has type $t2. r=bsmedberg

+++ xpcom/components/nsNativeComponentLoader.cpp
@@ -173,7 +173,7 @@ nsNativeModuleLoader::LoadModule(nsILocalFile* aFile, nsIModule* *aResult)
         fprintf(stderr,
                 "nsNativeModuleLoader::LoadModule(\"%s\") - load FAILED, "
                 "rv: %lx, error:\n\t%s\n",
-                filePath.get(), rv, errorMsg);
+                filePath.get(), (unsigned long)rv, errorMsg);
 #endif
 
         return rv;
