+++ xpcom/typelib/xpt/tools/xpt_dump.c
@@ -227,7 +227,8 @@ main(int argc, char **argv)
     if (flen > 0) {
         size_t rv = fread(whole, 1, flen, in);
         if (rv < flen) {
-            fprintf(stderr, "short read (%d vs %d)! ouch!\n", rv, flen);
+            fprintf(stderr, "short read (%u vs %u)! ouch!\n",
+                    (unsigned int)rv, (unsigned int)flen);
             goto out;
         }
         if (ferror(in) != 0 || fclose(in) != 0)
